code/deployment/api/requirements.txt
#dvc==3.55.2

zenml[server]
pendulum
mlflow


gunicorn
uvicorn
fastapi
python-dotenv
python-dotenv==1.0.1

torch==2.2.2
numpy==1.26.4
pandas==2.2.2

